// src/shared/types.d.ts
  interface Commits {
    reponame: string;
    additions: number;
    deletions: number;
    total: number;
    message: string;
    time: number;
    committer: string;
    commit: string;
  }
interface oCommits {
  reponame: string;
  additions: number;
  deletions: number;
  total: number;
  message: string;
  time: string;
  committer: string;
  commit: string;
}